uint8_t update

+++ src/ax25_utils.h
@@ -40,7 +40,7 @@ namespace AX25_Utils {
     String AX25FrameToLoRaPacket(String frame);
     String frameCleaning(String frame);
     std::string intToBinaryString(int value, int bitLength);
-    String encodeAX25Address(String frame, int type, bool lastAddress);
+    String encodeAX25Address(String frame, uint8_t type, bool lastAddress);
     String LoRaPacketToAX25Frame(String packet);
 
 }
