+++ src/gps_utils.cpp
@@ -74,11 +74,11 @@ namespace GPS_Utils {
     }
 
     void calculateHeadingDelta(int speed) {
-        int TurnMinAngle;
+        uint8_t TurnMinAngle;
         double headingDelta = abs(previousHeading - currentHeading);
         if (lastTx > currentBeacon->minDeltaBeacon * 1000) {
             if (speed == 0) {
-                TurnMinAngle = currentBeacon->turnMinDeg + (currentBeacon->turnSlope/(speed+1));
+                TurnMinAngle = currentBeacon->turnMinDeg + (currentBeacon->turnSlope/(speed + 1));
             } else {
                 TurnMinAngle = currentBeacon->turnMinDeg + (currentBeacon->turnSlope/speed);
             }
